Merge SDH's GB Sounds
* Implement HQ sound mixer

* WIP progress of GB Sounds engine

The GB Sounds engine has been ported and mostly works with the HQ mixer. I'm most likely missing some stuff as square and wave data aren't updating properly when swapping from a GB Sounds straight to m4a track and vice versa.

That aside, all songs from GSC are in the repo and one song from RBY has been ported to work with the engine (need to see if it sounds like it did in RBY)

* Bit of clean up

* More clean up

* Added RBY's trainer battle music

I have also split the music into their own folders based on the games they originated from.

An "examples" folder has also been added which contains music that have been ported but aren't used and is only there for if people want to use any of the music in their hacks.

* Add more GBS tracks and reorganise things a bit

Added some more tracks for the GBS engine and reorganised things so the drum noises have been given names and the note frequency table, wave samples and drum noise data have been put into their own files similar to pokecrystal.

* Made GB Sounds less invasive

Thanks to AsparagusEduardo!

* Update mixer code

* Fix spacing in mixer code

* Organise things a little better

Removed the "rby", "gsc" and "rse" folders and renamed the songs to match the normal versions.

* Route 101 and Abandoned Ship

* WIP refactor of songs and engine

Music for the GBS are now able to use song headers that match pokecrystal with the added option for priority. The GBS engine has been given numerous optimisations and functions have been moved so it's less invasive.

* Fix build error

* Add pitch slide for channels 2 & 3

Allows channel 2 and the wave channel on the GBS to use pitch_slide which allows RBY's titlescreen to play correctly.

* updated upgrade script

* added battle_sim.s as sample; updated vast_poni_canyon.s

* added alola_wild_battle.s; small fix to upgrade script

---------

Co-authored-by: Josh <32826900+ShinyDragonHunter@users.noreply.github.com>